Create missingElement.js
You have given an  array which contains numbers but one number is missing, you need to write a  program to find that missing number in an array. One trick to solve this problem is to calculate sum of all numbers in the array and compare with expected sum, the difference would be the missing number.