Rename anagramCheck.js to AnagramCheck.js
Title: check if two Strings are anagrams of each other

Body:

A simple coding problem based upon String, but could also be asked with numbers. You need to write a program to check if two given strings are anagrams of Each other. Two strings are anagrams if they are written using the same exact letters, ignoring space, punctuation and capitalization. Each letter should have the same count in both strings. For example, Army and Mary are anagram of each other.